calc-be/requirements.txt
# Core FastAPI dependencies
fastapi==0.112.2
uvicorn[standard]==0.30.6
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0

# Google AI dependencies - optimized versions
google-generativeai==0.7.2
grpcio>=1.60.0,<1.67.0  # Use range instead of exact version
grpcio-status>=1.60.0,<1.63.0
google-ai-generativelanguage==0.6.6
google-api-core[grpc]==2.19.2
google-auth==2.34.0
googleapis-common-protos==1.65.0
proto-plus==1.24.0
protobuf==4.25.4

# Other dependencies
pydantic==2.8.2
starlette==0.38.2
anyio==4.4.0
click==8.1.7
h11==0.14.0
sniffio==1.3.1
typing_extensions==4.12.2
idna==3.8
certifi==2024.8.30
charset-normalizer==3.3.2
urllib3==2.2.2
